rename sonar-dat to sonar-core
