+++ sonar-core/LICENSE
