+++ sonar-core/index.js
