+++ sonar-core/lib/config.js
