+++ sonar-core/lib/fs.js
