+++ sonar-core/lib/island.js
