+++ sonar-core/lib/schemas.js
