+++ sonar-core/lib/search/index-manager.js
