+++ sonar-core/lib/search/query.js
