+++ sonar-core/lib/search/schema.js
