+++ sonar-core/lib/search/view-sonar.js
