+++ sonar-core/test/basic.js
