+++ sonar-core/test/lib/example-schemas.js
