+++ sonar-core/test/lib/util.js
