+++ sonar-core/test/search-schema.js
