+++ sonar-server/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@arso-project/sonar-cli": "^0.2.2",
     "@arso-project/sonar-client": "^0.2.2",
-    "@arso-project/sonar-dat": "^0.2.0",
+    "@arso-project/sonar-core": "^0.2.0",
     "@arso-project/sonar-ui": "^0.2.2",
     "async-exit-hook": "^2.0.1",
     "body-parser": "^1.19.0",
